Create sport controller and service
*refact sport by gender
